bblfshctl: add an option to use v2 parsing protocol
